simple-definition-list-blocks/div の構成オブジェクトを修正
