+++ src/blocks/definition-div/block.json
@@ -2,9 +2,9 @@
   "$schema": "https://schemas.wp.org/trunk/block.json",
   "apiVersion": 3,
   "name": "simple-definition-list-blocks/div",
-  "title": "div tag (Simple Definition List Blocks)",
-  "icon": "welcome-write-blog",
-  "description": "Create div html tag.",
+  "title": "Div tag",
+  "icon": "title",
+  "description": "Add a div tag for inside the dl tag.",
   "textdomain": "simple-definition-list-blocks",
   "category": "text",
   "parent": [ "simple-definition-list-blocks/list" ],
@@ -15,6 +15,6 @@
       "selector": "div.wp-block-simple-definition-list-blocks-div"
     }
   },
-  "keywords": [ "definition", "div" ],
-  "editorScript": "file:../../../build/index.js"
+  "keywords": [ "definition list", "div" ],
+  "editorScript": "file:./index.js"
 }
