+++ src/blocks/definition-div/edit.js
@@ -1,0 +1,21 @@
+import {
+	InnerBlocks,
+	useBlockProps,
+	useInnerBlocksProps,
+} from '@wordpress/block-editor';
+
+const ALLOWED_BLOCKS = [
+	'simple-definition-list-blocks/term',
+	'simple-definition-list-blocks/details',
+	'simple-definition-list-blocks/details-html',
+];
+
+export default function Edit() {
+	const blockProps = useBlockProps();
+	const innerBlocksProps = useInnerBlocksProps( blockProps, {
+		allowedBlocks: ALLOWED_BLOCKS,
+		renderAppender: InnerBlocks.ButtonBlockAppender,
+	} );
+
+	return <div { ...innerBlocksProps } />;
+}
