+++ src/blocks/definition-div/index.js
@@ -1,32 +1,35 @@
 import { registerBlockType } from '@wordpress/blocks';
-import { InnerBlocks, useBlockProps, useInnerBlocksProps } from '@wordpress/block-editor';
+import { title, Icon } from '@wordpress/icons';
+import { __ } from '@wordpress/i18n';
 import metadata from './block.json';
 
-const ALLOWED_BLOCKS = [
-	'simple-definition-list-blocks/term',
-	'simple-definition-list-blocks/details',
-	'simple-definition-list-blocks/details-html',
-];
+/**
+ * Internal dependencies
+ */
+import Edit from './edit';
+import Save from './save';
 
-registerBlockType( metadata, {
-	...metadata,
-	edit: () => {
-		const blockProps = useBlockProps();
-		const innerBlocksProps = useInnerBlocksProps( blockProps, {
-			allowedBlocks: ALLOWED_BLOCKS,
-			renderAppender: InnerBlocks.ButtonBlockAppender
-		} );
+const { name, attributes, category, parent } = metadata;
 
-		return (
-			<div { ...innerBlocksProps } />
-		);
-	},
-	save: () => {
-		const saveBlockProps = useBlockProps.save();
-		const saveInnerBlocksProps = useInnerBlocksProps.save( saveBlockProps );
-
-		return (
-			<div { ...saveInnerBlocksProps } />
-		);
-	},
+registerBlockType( name, {
+	title: `${ __(
+		'Div tag',
+		'simple-definition-list-blocks'
+	) } (Simple Definition List Blocks)`,
+	description: __(
+		'Add a div tag for inside the dl tag.',
+		'simple-definition-list-blocks'
+	),
+	category,
+	icon: <Icon icon={ title } />,
+	keywords: [
+		'definition list',
+		__( 'definition list', 'simple-definition-list-blocks' ),
+		'div',
+		__( 'div', 'simple-definition-list-blocks' ),
+	],
+	parent,
+	attributes,
+	edit: Edit,
+	save: Save,
 } );
